The Abyss by Stanislas Marçais
Few updates like a new camera
Mouse wheel to zoom in the scene.
UP and DOWN to change the camera target creature.
mouse RIGHT to move the eye camera.
mouse LEFT   to move the scene position.

IN THE CODE
the colorMode now is HSB(360,100,100) ;
I change the background function for something easy for me to change
the color of the background.
I give a new var for the void creatureManager.loop() to give the color
for the creature.